Say goodbye to connection tracking in SemanticCore
Connection tracking was introduced while moving from rdf2go to sesame/rdf4j. With rdf2go, connections were not really necessary and we made sure, that all code refactored to sesame/rdf4j correctly handled the new connections. This is no longer necessary, since the code is relatively mature now. Additionally, with rdf4j 3, it seems the connection is quite the heavy weight, so we should not keep track of a large number of them due to memory issues.